+++ core/src/main/resources/reference.conf
@@ -114,7 +114,7 @@ jdbc-journal {
 
   tables {
     journal {
-      tableName = "journal"
+      tableName = "journal_event"
       schemaName = ""
 
       # Only used in pre 5.0.0 Dao. Renaming individual column names is
@@ -365,7 +365,7 @@ jdbc-read-journal {
   # to the same value for these other journals.
   use-shared-db = null
 
-  dao = "akka.persistence.jdbc.query.dao.ByteArrayReadJournalDao"
+  dao = "akka.persistence.jdbc.query.dao.AkkaSerializerReadJournalDao"
 
   # if true, queries will include logically deleted events
   # should not be configured directly, but through property akka-persistence-jdbc.logicalDelete.enable
@@ -392,7 +392,7 @@ jdbc-read-journal {
 
   tables {
     journal {
-      tableName = "journal"
+      tableName = "journal_event"
       schemaName = ""
       columnNames {
         ordering = "ordering"
