+++ core/src/main/scala/akka/persistence/jdbc/query/dao/legacy/ReadJournalQueries.scala
@@ -1,8 +1,12 @@
-package akka.persistence.jdbc.query.dao
+/*
+ * Copyright (C) 2014 - 2019 Dennis Vriend <https://github.com/dnvriend>
+ * Copyright (C) 2019 - 2020 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.persistence.jdbc.query.dao.legacy
 
 import akka.persistence.jdbc.config.{ JournalTableConfiguration, ReadJournalConfig }
-import akka.persistence.jdbc.journal.dao.JournalTables
-import akka.persistence.jdbc.journal.dao.legacy.JournalRow
+import akka.persistence.jdbc.journal.dao.legacy.{ JournalRow, JournalTables }
 import slick.jdbc.JdbcProfile
 
 class ReadJournalQueries(val profile: JdbcProfile, val readJournalConfig: ReadJournalConfig) extends JournalTables {
@@ -22,10 +26,6 @@ class ReadJournalQueries(val profile: JdbcProfile, val readJournalConfig: ReadJo
   private def baseTableQuery() =
     if (readJournalConfig.includeDeleted) JournalTable
     else JournalTable.filter(_.deleted === false)
-
-  private def baseTableWithTagsQuery() = {
-    baseTableQuery().join(TagTable).on(_.ordering === _.eventId)
-  }
 
   val allPersistenceIdsDistinct = Compiled(_allPersistenceIdsDistinct _)
 
@@ -48,12 +48,11 @@ class ReadJournalQueries(val profile: JdbcProfile, val readJournalConfig: ReadJo
       offset: ConstColumn[Long],
       maxOffset: ConstColumn[Long],
       max: ConstColumn[Long]) = {
-    baseTableWithTagsQuery()
-      .filter(_._2.tag === tag)
-      .sortBy(_._1.ordering.asc)
-      .filter(row => row._1.ordering > offset && row._1.ordering <= maxOffset)
+    baseTableQuery()
+      .filter(_.tags.like(tag))
+      .sortBy(_.ordering.asc)
+      .filter(row => row.ordering > offset && row.ordering <= maxOffset)
       .take(max)
-      .map(_._1)
   }
 
   val eventsByTag = Compiled(_eventsByTag _)
