+++ core/src/test/resources/schema/postgres/postgres-schema-v5.sql
@@ -1,12 +1,14 @@
-DROP TABLE IF EXISTS public.journal;
+DROP TABLE IF EXISTS public.event_tag;
+DROP TABLE IF EXISTS public.journal_event;
 
-CREATE TABLE IF NOT EXISTS public.journal (
+CREATE TABLE IF NOT EXISTS public.journal_event (
   ordering BIGSERIAL,
   persistence_id VARCHAR(255) NOT NULL,
   sequence_number BIGINT NOT NULL,
   deleted BOOLEAN DEFAULT FALSE NOT NULL,
 
   writer VARCHAR(255) NOT NULL,
+  write_timestamp BIGINT,
   event_manifest VARCHAR(255),
 
   event_ser_id INTEGER NOT NULL,
@@ -20,7 +22,16 @@ CREATE TABLE IF NOT EXISTS public.journal (
   PRIMARY KEY(persistence_id, sequence_number)
 );
 
-CREATE UNIQUE INDEX journal_ordering_idx ON public.journal(ordering);
+CREATE UNIQUE INDEX journal_event_ordering_idx ON public.journal_event(ordering);
+
+CREATE TABLE IF NOT EXISTS public.event_tag(
+    event_id BIGINT,
+    tag VARCHAR(256),
+    CONSTRAINT fk_journal_event
+      FOREIGN KEY(event_id)
+	  REFERENCES journal_event(ordering)
+	  ON DELETE CASCADE
+);
 
 DROP TABLE IF EXISTS public.snapshot;
